[feat] Add api-server for user to retrieve SignerKey
Background of this api-server implementation is a simple permission
issue. When a user have get/list permission for SignerKey, there is a
problem which he/she can see other users' SignerKey. If the user does
not have get/list permission, user cannot use the key personally, e.g.,
if he/she wants to sign/push the image locally.

Simple idea here is to use api-server. A user who is granted to use an
ImageSigner, he is permitted for a subresource /keys for the resource.
The endpoint for the subresource is like
'/apis/registry.tmax.io/v1/imagesigners/<ImageSigner name>/keys'.
And, importantly, no user is permitted to get/list SignerKey resources.
This can solve the security issue.

Role of this api-server is simple, it only validates the user and
returns the complete body of SignerKey when the api endpoint is called.